requirements.txt
python-telegram-bot==20.7
openai==1.35.13
httpx==0.27.0